chore(): separa os arquivos de PictureDetail
